Start inter-bot code
