+++ inter-page/support-bot/app.js
@@ -1,0 +1,154 @@
+/**
+ * Copyright 2017-present, Facebook, Inc. All rights reserved.
+ *
+ * This source code is licensed under the license found in the
+ * LICENSE file in the root directory of this source tree.
+ *
+ *
+ * Starter Project for Messenger Platform Quick Start Tutorial
+ *
+ * Use this project as the starting point for following the
+ * Messenger Platform quick start tutorial.
+ *
+ * https://developers.facebook.com/docs/messenger-platform/getting-started/quick-start/
+ *
+ */
+
+'use strict';
+
+// Imports dependencies and set up http server
+const
+    request = require('request'),
+    express = require('express'),
+    body_parser = require('body-parser'),
+    app = express().use(body_parser.json()); // creates express http server
+
+// Sets server port and logs message on success
+app.listen(process.env.PORT || 5001, () => console.log('webhook is listening'));
+
+app.post('/webhook', (req, res) => {
+
+    // Parse the request body from the POST
+    let body = req.body;
+
+    // Check the webhook event is from a Page subscription
+    if (body.object === 'page') {
+
+        body.entry.forEach(function (entry) {
+
+            // Gets the body of the webhook event
+            let webhook_event = entry.messaging[0];
+            console.log(webhook_event);
+
+            // Get the sender PSID
+            let sender_psid = webhook_event.sender.id;
+            console.log('Sender PSID: ' + sender_psid);
+
+            // Check if the event is a message or postback and
+            // pass the event to the appropriate handler function
+            if (webhook_event.message) {
+                handleMessage(sender_psid, webhook_event.message);
+            } else if (webhook_event.postback) {
+                handleIncoming(sender_psid, webhook_event.postback);
+            } else if (webhook_event.referral) {
+                handleIncoming(sender_psid, webhook_event.referral);
+            }
+
+        });
+
+        // Return a '200 OK' response to all events
+        res.status(200).send('EVENT_RECEIVED');
+
+    } else {
+        // Return a '404 Not Found' if event is not from a page subscription
+        res.sendStatus(404);
+    }
+
+});
+
+// Accepts GET requests at the /webhook endpoint
+app.get('/webhook', (req, res) => {
+
+    const VERIFY_TOKEN = "TOKEN";
+
+    // Parse params from the webhook verification request
+    let mode = req.query['hub.mode'];
+    let token = req.query['hub.verify_token'];
+    let challenge = req.query['hub.challenge'];
+
+    // Check if a token and mode were sent
+    if (mode && token) {
+
+        // Check the mode and token sent are correct
+        if (mode === 'subscribe' && token === VERIFY_TOKEN) {
+
+            // Respond with 200 OK and challenge token from the request
+            console.log('WEBHOOK_VERIFIED');
+            res.status(200).send(challenge);
+
+        } else {
+            // Responds with '403 Forbidden' if verify tokens do not match
+            res.sendStatus(403);
+        }
+    }
+});
+
+
+function handleMessage(sender_psid, received_message) {
+    let response;
+
+    if (received_message.text) {
+        switch (received_message.text.replace(/[^\w\s]/gi, '').trim().toLowerCase()) {
+            default:
+                response = {
+                    "text": `You sent the message: "${received_message.text}".`
+                };
+                break;
+        }
+    } else {
+        response = {
+            "text": `Sorry, I don't understand what you mean.`
+        }
+    }
+
+    // Send the response message
+    callSendAPI(sender_psid, response);
+}
+
+function handleIncoming(sender_psid, received_postback) {
+    let response;
+    if (received_postback.ref) {
+        response = {"text": `Nice to see you again. I have your original message, "${received_postback.ref}". \n Let me see how I can help.`}
+    } else {
+        let payload = received_postback.payload;
+        if (payload === 'get_started') {
+            response = {"text": `Welcome. I have your original message, "${received_postback.referral.ref}". \n Let me see how I can help.`}
+        }
+    }
+    // Send the message to acknowledge the postback
+    callSendAPI(sender_psid, response);
+}
+
+function callSendAPI(sender_psid, response) {
+    // Construct the message body
+    let request_body = {
+        "recipient": {
+            "id": sender_psid
+        },
+        "message": response
+    };
+
+    // Send the HTTP request to the Messenger Platform
+    request({
+        "uri": "https://graph.facebook.com/v2.6/me/messages",
+        "qs": {"access_token": "EAACOkZC1Y5pABAAHOrScsZAI5ZBFf4cj5MwzE62eWX67y1YcNpZBhrWSvcq6elZCH4Nh9BZAvzESYfNQqkTXD2V1ZBljbTzVO4ZBpFFMc8XlFuVuQ82ZCteO5DG7TDLOfBCRKZAfJ89IfZC4qNyPqAzvuyR19pdE5DRptfY7hULt26BTQZDZD"},
+        "method": "POST",
+        "json": request_body
+    }, (err, res, body) => {
+        if (!err) {
+            console.log('message sent!')
+        } else {
+            console.error("Unable to send message:" + err);
+        }
+    });
+}
