+++ inter-page/support-bot/package.json
@@ -15,7 +15,6 @@
   "dependencies": {
     "body-parser": "^1.15.0",
     "express": "^4.13.4",
-    "fb": "^2.0.0",
     "request": "^2.72.0"
   },
   "engines": {
